+++ src/js/Level.class.js
@@ -1,4 +1,4 @@
-import { levelEvents, movements } from "./utils";
+import { levelEvents } from "./utils";
 
 export class Level extends EventTarget {
   #ctx;
@@ -6,11 +6,10 @@ export class Level extends EventTarget {
   #renderer;
   #eventHandler;
 
-  #abortController = new AbortController();
   #key = {};
-  doorPosition = Math.floor(Math.random() * 16);
-  playerPosition = Math.floor(Math.random() * 16);
-  keyPosition = Math.floor(Math.random() * 16);
+  #doorPosition = Math.floor(Math.random() * 16);
+  #playerPosition = Math.floor(Math.random() * 16);
+  #keyPosition = Math.floor(Math.random() * 16);
 
   constructor({ ctx, player, eventHandler, renderer }) {
     super();
@@ -22,48 +21,43 @@ export class Level extends EventTarget {
 
     this.#renderer.setColor(this);
 
-    this.#eventHandler.addEventListener(movements.up, () => this.#up(), {
-      signal: this.#abortController.signal,
-    });
-    this.#eventHandler.addEventListener(movements.down, () => this.#down(), {
-      signal: this.#abortController.signal,
-    });
-    this.#eventHandler.addEventListener(movements.left, () => this.#left(), {
-      signal: this.#abortController.signal,
-    });
-    this.#eventHandler.addEventListener(movements.right, () => this.#right(), {
-      signal: this.#abortController.signal,
-    });
-
     this.#render();
   }
 
-  #up() {
-    this.playerPosition = (16 + this.playerPosition - 4) % 16;
+  up() {
+    this.#playerPosition = (16 + this.#playerPosition - 4) % 16;
     this.#render();
   }
 
-  #down() {
-    this.playerPosition = (this.playerPosition + 4) % 16;
+  down() {
+    this.#playerPosition = (this.#playerPosition + 4) % 16;
     this.#render();
   }
 
-  #left() {
-    const row = Math.floor(this.playerPosition / 4);
-    this.playerPosition = 4 * row + ((4 + this.playerPosition - 1) % 4);
+  left() {
+    const row = Math.floor(this.#playerPosition / 4);
+    this.#playerPosition = 4 * row + ((4 + this.#playerPosition - 1) % 4);
     this.#render();
   }
 
-  #right() {
-    const row = Math.floor(this.playerPosition / 4);
-    this.playerPosition = 4 * row + ((this.playerPosition + 1) % 4);
+  right() {
+    const row = Math.floor(this.#playerPosition / 4);
+    this.#playerPosition = 4 * row + ((this.#playerPosition + 1) % 4);
     this.#render();
+  }
+
+  get position() {
+    return {
+      door: this.#doorPosition,
+      key: this.#keyPosition,
+      player: this.#playerPosition,
+    };
   }
 
   #preRenderActions() {
     if (this.#player.has(this.#key)) {
-      this.keyPosition = this.playerPosition;
-    } else if (this.playerPosition === this.keyPosition) {
+      this.#keyPosition = this.#playerPosition;
+    } else if (this.#playerPosition === this.#keyPosition) {
       this.#player.key = this.#key;
     }
 
@@ -72,7 +66,7 @@ export class Level extends EventTarget {
 
   #afterRenderActions() {
     if (
-      this.doorPosition === this.playerPosition &&
+      this.#doorPosition === this.#playerPosition &&
       this.#player.has(this.#key)
     ) {
       return this.#onDone();
@@ -88,8 +82,6 @@ export class Level extends EventTarget {
   }
 
   #onDone() {
-    this.#abortController.abort();
-
     this.dispatchEvent(new Event(levelEvents.done));
   }
 }